requirements.txt
absl-py==1.4.0
torch-geometric==2.3.1
scikit-learn==1.2.2
scipy==1.10.0
pytorch-lightning==1.9.4
networkx==3.1
tensorboard==2.13.0